updates in desenvolvimento tab

+++ public/ajax/cache/update.php
@@ -1,4 +1,10 @@
 <?php
 
-unlink(PATH_HOME . "_config/updates/version.txt");
-$up = new \Dashboard\UpdateDashboard();
+try {
+    if (file_exists(PATH_HOME . "_config/updates/version.txt"))
+        unlink(PATH_HOME . "_config/updates/version.txt");
+
+    $up = new \Dashboard\UpdateDashboard([]);
+} catch (Exception $e) {
+
+}
